+++ src/encode/error.rs
@@ -17,8 +17,8 @@ pub enum Error {
     InvalidChunkSequence,
     /// Chunk is too large to save in a PNG file (length must fit in 32 bits)
     ChunkTooBig,
-    /// Text is not between 1-79 characters
-    TextSize(usize),
+    /// key is not between 1-79 characters
+    KeySize(usize),
     /// PLTE chunk with a palette that has less than 1 or more than 256 colors
     BadPalette,
     /// Chunks arranged in invalid sequence.  Provides PNG chunk identifier of
@@ -33,11 +33,9 @@ impl std::fmt::Display for Error {
             Io(io) => write!(f, "I/O Error: {}", io),
             InvalidChunkSequence => write!(f, "Invalid chunk sequence"),
             ChunkTooBig => write!(f, "Chunk too big"),
-            TextSize(size) => write!(
-                f,
-                "Text size {} is not between 1 and 79 characters",
-                size
-            ),
+            KeySize(size) => {
+                write!(f, "Key size {size} is not between 1 and 79 characters")
+            }
             BadPalette => write!(f, "Invalid palette"),
             ChunkOrder(bytes) => write!(
                 f,